pyproject.toml
[project]
name = "ontoeval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Chris Mungall", email = "cjm@berkeleybop.org" }
]
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.8",
    "diskcache>=5.6.3",
    "joblib>=1.5.1",
    "pandas>=2.3.0",
    "pydantic>=2.11.5",
    "pydantic-ai>=0.2.15",
    "pyyaml>=6.0.2",
]

[project.scripts]
ontoeval = "ontoeval:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]

[tool.pytest.ini_options]
addopts = "--doctest-modules"
doctest_optionflags = ["ELLIPSIS", "NORMALIZE_WHITESPACE"]
